pyproject.toml
[tool.poetry]
name = "ggj-2024"
version = "0.1.0"
description = ""
authors = ["Jacob Cheatley <jacobcheatley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
websockets = "^12.0"
requests = "^2.31.0"
uvicorn = "^0.27.0"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
numpy = "^1.26.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120